Implemented functional menus & deconstructed giant javascript file.
