+++ src/content/events.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+const events = 
+<div>
+<h1>Right Now</h1>
+<iframe href="foo"></iframe>
+<p>
+I am a placeholder for a Twitch steam of a tournament going on right now. I know I am going on right now because the Google Calandar says so, and I know what stream to show because they put it in the description. If there are no Google Calandar events going on right now, I will tell the user there are no tournaments happing right now.
+</p>
+
+<h1>Next</h1>
+<iframe href="foo"></iframe>
+<p>
+I am a placeholder for our Google Calandar. By having a Google Calandar instead of a database to know upcoming tournaments, tournament hosts promoting their tournaments don't need any keys or knowlege on our website, and we don't need to do any mainteinance or changes to the website for them.
+</p>
+
+<p>How do I put my tournament on here?<br />
+Ask our project manager for access (see footer for social media links). They will fill you in on our format and guidelines.</p>
+</div>
+
+export default events;
