+++ src/content/facebook.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Popup from '../class/Popup.js';
+import SocialMedia from '../class/SocialMedia.js';
+
+const facebook = 
+<Popup title="facebook" children={
+<p>The lead developer is doesn't use facebook and the others are all shy and didn't link their facebook accounts to their Discord profiles. uwu</p>}/>
+
+export default facebook;
