+++ src/content/gettingStarted.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+const gettingStarted = 
+<div>
+<h1>Welcome!</h1>
+<h2>Rulesets</h2>
+<p>Rulesets are a per-tournament, self-imposed set of rules to promote competition and fairness. Usually Smash tournaments rulesets restrict the following.</p>
+
+<p>Stages you can pick<br />
+Many stages have walk-offs which lead to degenerate play, and some platforms and shapes can mess with core moves like recoveries. For example, many characters with teleporting recoveries will have their recovery send them in the wrong direction because of a slant under the stage, when they otherwise would have made it back.</p>
+
+<p>The order you can pick stages<br />
+Even if we restrict the stages we can play, there's bound to be stages that favor some characters over others. In the spirit of fair competition, stage selection should not determine who wins. Many rulesets will not let the winner pick the next stage to prevent snowballing.</p>
+
+<p>Stock count (usually three stocks)<br />
+Between the many stocks of Melee due to its fast pace, and the few stocks of Smash Bros. for Wii U due to its slow pace, competitors have decided to settle in the middle for Smash Ultimate. Using stocks instead of time makes the playing field more level between characters who kill fast and slow, increasing the variety of characters viable for competition.</p>
+
+<p>Items<br />
+Items easily lead to snowballing.</p>
+
+<h2>Code of Conduct</h2>
+<p>Tournaments usually have an entry fee which pays for the building, staffing, and winners' prizes. If you're clean, aren't an ass, and bring your own controller and you'll usually be fine. While matches are going on, if you're planning on using that free set-up over there, ask first, as it's probably reserved for an upcoming match. Usually, staff you can ask questions and report match results to are on a computer or walking around announcing the next match. At the end of each set, one person needs to tell the staff the match results. Your safest bet is the one on the computer.</p>
+
+<h2>Entering a tournament!</h2>
+<p>If you've never played competitive smash before and want to try it out, try going to a 'local'! A local is the smallest type of tournament that registers on leaderboards. Because of their small size, usually at locals it's easy to make friends, and they're relatively cheap and short. Go to 'Events' to find a local right for you!</p>
+</div>
+
+export default gettingStarted;
