+++ src/content/github.js
@@ -1,0 +1,46 @@
+import React from 'react';
+import Popup from '../class/Popup.js';
+import SocialMedia from '../class/SocialMedia.js';
+
+const github = 
+<Popup title="GitHub" children={
+<div>
+<SocialMedia
+	profileImg="https://avatars1.githubusercontent.com/u/54699262?s=200&v=4"
+  role="Repo"
+	url="https://github.com/OregonSmashUltimate/OregonSmashUltimate.github.io"
+	username="OregonSmashUltimate.github.io"/>
+<hr />
+<SocialMedia
+	profileImg="https://avatars0.githubusercontent.com/u/10984092?s=460&v=4"
+  role="Project Manager"
+	url="https://github.com/malin383"
+	username="malin383"/>
+<hr />
+<SocialMedia
+	profileImg="https://avatars1.githubusercontent.com/u/1504375?s=460&v=4"
+  role="Developer Lead"
+	url="https://github.com/cyril-vandyke"
+	username="cyril-vandyke"/>
+<hr />
+<SocialMedia
+	profileImg="https://avatars3.githubusercontent.com/u/21343049?s=460&v=4"
+  role="Lead Developer"
+	url="https://github.com/jCallon"
+	username="jCallon"/>
+<hr />
+<SocialMedia
+	profileImg="https://avatars3.githubusercontent.com/u/253703?s=460&v=4"
+  role="Human Database"
+	url="https://github.com/giantrobotbee"
+	username="giantrobotbee"/>
+<hr />
+<SocialMedia
+	profileImg="https://avatars2.githubusercontent.com/u/29513096?s=460&v=4"
+  role="Logo Design"
+	url="https://github.com/bananyog"
+	username="bananyog"/>
+</div>
+}/>
+
+export default github;
