+++ src/content/rankings.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const rankings = 
+<div>
+<h1>Rankings are kept track of on <a href="https://smash.gg/rankings/" target="_blank" rel="noopener noreferrer">smash.gg</a></h1>
+<p>When looking up rankings, keep in mind they are separated by season, so they are not always up to date. Rankings are decided by points earned from tournament placings.</p>
+</div>
+
+export default rankings;
