+++ src/content/twitter.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Popup from '../class/Popup.js';
+import SocialMedia from '../class/SocialMedia.js';
+
+const twitter = 
+<Popup title="Twitter" children={
+<SocialMedia
+	profileImg="https://pbs.twimg.com/profile_images/1120052132776796160/nbK9A9G7_400x400.jpg"
+  role="Lead Developer"
+	url="https://twitter.com/CeruleanAoi?s=09"
+	username="@CeruleanAoi"
+/>}/>
+
+export default twitter;
